+++ platformio.ini
@@ -1,0 +1,15 @@
+; PlatformIO Project Configuration File
+;
+;   Build options: build flags, source filter
+;   Upload options: custom upload port, speed and extra flags
+;   Library options: dependencies, extra library storages
+;   Advanced options: extra scripting
+;
+; Please visit documentation for the other options and examples
+; https://docs.platformio.org/page/projectconf.html
+
+[env:sparkfun_promicro16]
+platform = atmelavr
+board = sparkfun_promicro16
+framework = arduino
+lib_deps = dmadison/XInput@^1.2.6
